requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=0.15.0=pypi_0
accelerate=1.0.1=pypi_0
addict=2.4.0=pypi_0
aiohappyeyeballs=2.4.0=pypi_0
aiohttp=3.10.5=pypi_0
aioice=0.9.0=pypi_0
aiortc=1.9.0=pypi_0
aiosignal=1.3.1=pypi_0
albucore=0.0.17=pypi_0
albumentations=1.4.18=pypi_0
altair=4.2.2=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
av=12.3.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports-zoneinfo=0.2.1=pypi_0
basicsr=1.4.2=pypi_0
beautifulsoup4=4.11.2=pypi_0
bitsandbytes=0.43.3=pypi_0
blinker=1.8.2=pypi_0
blobfile=3.0.0=pypi_0
brisque=0.0.16=pypi_0
ca-certificates=2024.7.2=h06a4308_0
cachetools=5.5.0=pypi_0
certifi=2024.7.4=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.3.2=pypi_0
clarabel=0.9.0=pypi_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
cmake=3.30.2=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pyhd8ed1ab_0
commonmark=0.9.1=pypi_0
contourpy=1.1.1=pypi_0
cryptography=43.0.1=pypi_0
cvxpy=1.5.2=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.11=pypi_0
datasets=2.10.1=pypi_0
debugpy=1.6.7=py38h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
deepface=0.0.79=dev_0
deepfloyd-if=1.0.2rc0=pypi_0
diffusers=0.20.0=pypi_0
dill=0.3.6=pypi_0
dlib=19.24.6=pypi_0
dnnlib=0.0.1=pypi_0
dnspython=2.6.1=pypi_0
docker-pycreds=0.4.0=pypi_0
easydict=1.13=pypi_0
ecos=2.0.14=pypi_0
einops=0.7.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
eval-type-backport=0.2.2=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.0.1=pyhd8ed1ab_0
facexlib=0.3.0=pypi_0
fairscale=0.4.13=pypi_0
filelock=3.13.1=pypi_0
filterpy=1.4.5=pypi_0
fire=0.6.0=pypi_0
flask=3.0.3=pypi_0
flask-cors=4.0.2=pypi_0
flatbuffers=1.12=pypi_0
fonttools=4.53.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.6.1=pypi_0
ftfy=6.1.3=pypi_0
future=1.0.0=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gast=0.4.0=pypi_0
gdown=5.2.0=pypi_0
generic=1.1.3=pypi_0
geomloss=0.2.6=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth=2.34.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.70.0=pypi_0
gunicorn=23.0.0=pypi_0
h5py=3.1.0=pypi_0
huggingface-hub=0.25.2=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=pypi_0
ifaddr=0.2.0=pypi_0
image-quality=1.2.7=pypi_0
image-reward=1.5=pypi_0
imageio=2.35.1=pypi_0
importlib-metadata=8.4.0=pypi_0
importlib-resources=6.4.4=pypi_0
insightface=0.7.3=pypi_0
iopath=0.1.10=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.12.0=pyh41d4057_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=py38h578d9bd_0
kaleido=0.2.1=pypi_0
keras=2.13.1=pypi_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=18.1.1=pypi_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libsvm=3.23.0.4=pypi_0
libsvm-official=3.35.0=pypi_0
lightning-utilities=0.11.9=pypi_0
lit=18.1.8=pypi_0
llvmlite=0.41.1=pypi_0
lmdb=1.6.2=pypi_0
lpips=0.1.4=pypi_0
lxml=5.3.1=pypi_0
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mlp-mixer-pytorch=0.1.1=pypi_0
mpmath=1.3.0=pypi_0
mtcnn=0.1.1=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.14=pypi_0
mypy-extensions=1.0.0=pypi_0
narwhals=1.29.0=pypi_0
natsort=7.1.1=pyhd3eb1b0_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.0rc1=pypi_0
ninja=1.11.1.1=pypi_0
numba=0.58.1=pypi_0
numpy=1.23.4=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu11=2.21.5=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.20=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
onnx=1.17.0=pypi_0
onnxruntime-gpu=1.19.2=pypi_0
open-clip-torch=2.30.0=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openssl=3.0.14=h5eee18b_0
opt-einsum=3.3.0=pypi_0
osqp=0.6.7.post3=pypi_0
packaging=24.1=pyhd8ed1ab_0
pandas=1.4.4=pypi_0
parso=0.8.4=pyhd8ed1ab_0
peft=0.12.0=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pilgram=1.2.1=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py38h06a4308_0
piq=0.8.0=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.2.2=pyhd8ed1ab_0
plotly=6.0.0=pypi_0
portalocker=3.0.0=pypi_0
prettytable=3.11.0=pypi_0
prompt-toolkit=3.0.47=pyha770c72_0
prompt_toolkit=3.0.47=hd8ed1ab_0
protobuf=3.20.3=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyarrow=17.0.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.22=pypi_0
pycryptodomex=3.22.0=pypi_0
pydantic=2.10.4=pypi_0
pydantic-core=2.27.2=pypi_0
pydeck=0.9.1=pypi_0
pyee=12.0.0=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pylibsrtp=0.10.0=pypi_0
pympler=1.1=pypi_0
pynndescent=0.5.13=pypi_0
pyopenssl=24.2.1=pypi_0
pyparsing=3.1.4=pypi_0
pyre-extensions=0.0.23=pypi_0
pysocks=1.7.1=pypi_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.8=2_cp38
pytorch-lightning=2.0.3=pypi_0
pytorch-triton=3.0.0+dedb7bdf33=pypi_0
pytz=2024.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=25.1.2=py38h6a678d5_0
qdldl=0.1.7.post4=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.7.24=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
responses=0.18.0=pypi_0
retina-face=0.0.14=dev_0
rich=12.0.1=pypi_0
rpds-py=0.20.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.4=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
scs=3.2.7=pypi_0
seaborn=0.13.2=pypi_0
semver=3.0.2=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=2.13.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=72.1.0=py38h06a4308_0
six=1.15.0=pypi_0
smmap=5.0.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.10.0=pypi_0
streamlit-webrtc=0.42.0=pypi_0
sympy=1.12.1=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.5.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.6.13=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.20.3=pypi_0
toml=0.10.2=pypi_0
tomli=2.2.1=pypi_0
toolz=0.12.1=pypi_0
torch=2.4.1=pypi_0
torch-dct=0.1.6=pypi_0
torch-utils=0.1.2=pypi_0
torchmetrics=1.5.2=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.1=py38h0a891b7_3
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.46.3=pypi_0
triton=3.0.0=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.1=pypi_0
tzlocal=5.2=pypi_0
umap-learn=0.5.7=pypi_0
urllib3=2.2.2=pypi_0
validators=0.34.0=pypi_0
wandb=0.17.8=pypi_0
watchdog=4.0.2=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.4=pypi_0
wheel=0.43.0=py38h06a4308_0
wrapt=1.12.1=pypi_0
xformer=1.0.1=pypi_0
xformers=0.0.28.post1=pypi_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pypi_0
yapf=0.43.0=pypi_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.20.0=pypi_0
zlib=1.2.13=h5eee18b_1
